app/requirements.txt
sqlalchemy==2.0.30
flask==3.0.3
spacy==3.7.2
gunicorn==20.1.0
python-dotenv==1.0.1
pyspellchecker==0.8.1
requests==2.31.0
boto3==1.34.101
botocore==1.34.101
werkzeug==3.0.3
deepgram-sdk==3.2.7
flask-cors==4.0.1
stripe==9.5.0
Flask-SQLAlchemy==3.1.1